eliminando afecciones tabla por redundancia, funcionan en afecciones list
